kafka-structured/tests/requirements.txt
# Test dependencies for Kafka pipeline integration tests
kafka-python==2.0.2
